==> src/DILI-negative-oversampling/07_Obtain_feature_importances.ipynb <==
import os.path
%load_ext autoreload
%autoreload 2

import pandas as pd
from pandarallel import pandarallel
from tqdm.auto import tqdm, trange
from sklearn.metrics import auc as auc_fn
from sklearn.linear_model import LinearRegression

from utilities import deserialize_xgboost_classifier_and_get_importance_gain

Constants.

ROOT = '.'
OUT_FOLDER = f'{ROOT}/results'
DATA_FOLDER = f'{ROOT}/data'
FILE = f"{DATA_FOLDER}/Supp. File 1.xlsx"
N_JOBS = -1

Read results from the custom oversampling.

if not os.path.exists(f'{DATA_FOLDER}/DILI_complete_oversampling_analysis2024_feature_importance_avg_gain_norm_across_folds.feather'):
    data = pd.read_feather(f'{DATA_FOLDER}/DILI_complete_oversampling_analysis2024.feather')

Extract feature importances from serialized models.

if not os.path.exists(f'{DATA_FOLDER}/DILI_complete_oversampling_analysis2024_feature_importance_avg_gain_norm_across_folds.feather'):
    pandarallel.initialize(progress_bar=True, verbose=False)
    important_features = pd.DataFrame.from_dict(data['model'].parallel_apply(deserialize_xgboost_classifier_and_get_importance_gain).tolist())
    important_features.insert(0, 'dataset', data['dataset'].reset_index(drop=True), True)
    important_features.insert(1, 'descriptors', data['descriptors'].reset_index(drop=True), True)
    important_features.insert(2, 'num_stds', data['num_stds'].reset_index(drop=True), True)
    important_features.insert(3, 'samples added', data['samples added'].reset_index(drop=True), True)

if not os.path.exists(f'{DATA_FOLDER}/DILI_complete_oversampling_analysis2024_feature_importance_avg_gain_norm_across_folds.feather'):
    # Ensure total gain of all features for each condition = 1
    for i in trange(0, len(important_features), 1000):
        important_features.iloc[i: i +1000, 4:] = important_features.iloc[i: i +1000, 4:].div(important_features.iloc[i: i +1000, 4:].sum(axis=1), axis=0)
    
    important_features.to_feather(f'{OUT_FOLDER}/DILI_complete_oversampling_analysis2024_feature_importance_avg_gain_norm.feather')
    # Sum across 5-folds
    important_features = important_features.groupby(['dataset', 'descriptors', 'num_stds', 'samples added'], as_index=False).sum()
    # Ensure total gain is normalized per condition (their sums are 1)
    important_features.iloc[:, 4:] = important_features.iloc[:, 4:].div(important_features.iloc[:, 4:].sum(axis=1), axis=0)
    # Save to disk.
    important_features.to_feather(f'{DATA_FOLDER}/DILI_complete_oversampling_analysis2024_feature_importance_avg_gain_norm_across_folds.feather')

important_features = pd.read_feather(
    f'{DATA_FOLDER}/DILI_complete_oversampling_analysis2024_feature_importance_avg_gain_norm_across_folds.feather')
# Melt and ignore NAs summed to 0
data = important_features.melt(id_vars=['dataset', 'descriptors', 'num_stds', 'samples added']).query('value != 0')

Separate feature per type.

bio_descs = data[data.variable.str.split('_').str[0].isin(['bip', 'btg2', 'chop', 'hmox1', 'hspa1b', 'icam1', 'p21', 'srxn1'])].variable.drop_duplicates()
chem_descs = data[~data.variable.str.startswith('ECFP_6') & ~data.variable.str.split('_').str[0].isin(['bip', 'btg2', 'chop', 'hmox1', 'hspa1b', 'icam1', 'p21', 'srxn1']) & ~data.variable.str.endswith('_WT') & ~data.variable.str.contains('^.+(?:_[ACDEFGHIKLMNPQRSTVWY]\d+[ACDEFGHIKLMNPQRSTVWY])+$')].variable.drop_duplicates()
chemfp_descs = data[data.variable.str.startswith('ECFP_6')].variable.drop_duplicates()
allchem_descs = pd.concat((chem_descs, chemfp_descs)).drop_duplicates()
biospectra_descs = data[data.variable.str.endswith('_WT') & ~data.variable.str.contains('^.+(?:_[ACDEFGHIKLMNPQRSTVWY]\d+[ACDEFGHIKLMNPQRSTVWY])+$')].variable.drop_duplicates()

bip_descs = data[data.variable.str.startswith('bip')].variable.drop_duplicates()
btg2_descs = data[data.variable.str.startswith('btg2')].variable.drop_duplicates()
chop_descs = data[data.variable.str.startswith('chop')].variable.drop_duplicates()
hmox1_descs = data[data.variable.str.startswith('hmox1')].variable.drop_duplicates()
hspa1b_descs = data[data.variable.str.startswith('hspa1b')].variable.drop_duplicates()
icam1_descs = data[data.variable.str.startswith('icam1')].variable.drop_duplicates()
p21_descs = data[data.variable.str.startswith('p21')].variable.drop_duplicates()
srxn1_descs = data[data.variable.str.startswith('srxn1')].variable.drop_duplicates()

cmax1_descs = data[data.variable.str.split('_').str[1] == 'cmax1'].variable.drop_duplicates()
cmax5_descs = data[data.variable.str.split('_').str[1] == 'cmax5'].variable.drop_duplicates()
cmax10_descs = data[data.variable.str.split('_').str[1] == 'cmax10'].variable.drop_duplicates()
cmax25_descs = data[data.variable.str.split('_').str[1] == 'cmax25'].variable.drop_duplicates()
cmax50_descs = data[data.variable.str.split('_').str[1] == 'cmax50'].variable.drop_duplicates()
cmax100_descs = data[data.variable.str.split('_').str[1] == 'cmax100'].variable.drop_duplicates()

tp24 = data[data.variable.str.split('_').str[2] == 'tp24'].variable.drop_duplicates()
tp48 = data[data.variable.str.split('_').str[2] == 'tp48'].variable.drop_duplicates()
tp72 = data[data.variable.str.split('_').str[2] == 'tp72'].variable.drop_duplicates()

Obtain the feature importance per feature type.

features = [('bio', bio_descs), ('allchem', allchem_descs), ('chem', chem_descs), ('chemfp', chemfp_descs), ('biospectra', biospectra_descs),
            ('bip', bip_descs), ('btg2', btg2_descs), ('chop', chop_descs), ('hmox1', hmox1_descs), ('hspa1b', hspa1b_descs), ('icam1', icam1_descs), ('p21', p21_descs), ('srxn1', srxn1_descs),
            ('cmax1', cmax1_descs), ('cmax5', cmax5_descs), ('cmax10', cmax10_descs), ('cmax25', cmax25_descs), ('cmax50', cmax50_descs), ('cmax100', cmax100_descs),
            ('24h', tp24), ('48h', tp48), ('72h', tp72)]

table1 = []
for dataset in tqdm(['All', 'GFP', 'AnV', 'PI' , 'AnV-PI', 'GFP-AnV-PI',
                     'IntegratedIntensity', 'MeanIntensity', 'NormCounts']):
    for descriptors in ['bio', 'biochem', 'biospectra']:
        for num_stds in range(1, 6):
            x = data.query(f'dataset == "{dataset}" and descriptors == "{descriptors}" and num_stds == {num_stds}', engine='python')
            results1 = {}
            results2 = {}
            for feature_name, feature_list in features:
                values = x.query(f'variable in @feature_list', engine='python').groupby('samples added', as_index=False).sum()[['samples added', 'value']]
                if values.empty:
                    results1[feature_name] = None
                    results2[feature_name] = None
                else:
                    auc = auc_fn(values['samples added']/250, values['value'])
                    slope = LinearRegression().fit(values['samples added'].values.reshape(-1, 1), values['value']).coef_.item()
                    results1[feature_name] = auc
                    results2[feature_name] = slope
            results1 = pd.DataFrame([results1])
            results1.insert(0, 'dataset', dataset)
            results1.insert(1, 'descriptors', descriptors)
            results1.insert(2, 'num_stds', num_stds)
            results2 = pd.DataFrame([results2])
            results2.insert(0, 'dataset', dataset)
            results2.insert(1, 'descriptors', descriptors)
            results2.insert(2, 'num_stds', num_stds)
            table1.append(results1)

table1 = pd.concat(table1)

Obtain Figure 4A, B, and C.

table1.query('dataset == "All" and num_stds == 1').drop(columns=['dataset', 'num_stds']).groupby('descriptors').agg('mean').round(4).T

Obtain first part of Figure 4D. 

table3 = []
dataset  = 'All'
for descriptors in tqdm(['bio', 'biochem', 'biospectra']):
    for num_stds in range(1, 6):
        x = data.query(f'dataset == "{dataset}" and descriptors == "{descriptors}" and num_stds == {num_stds}', engine='python')
        results1 = {}
        results2 = {}
        for feature_name in chem_descs:
            values = x.query(f'variable == "{feature_name}"', engine='python')[['samples added', 'value']]
            if values.empty or values.shape[0] == 1:
                results1[feature_name] = None
            else:
                auc = auc_fn(values['samples added']/250, values['value'])
                slope = LinearRegression().fit(values['samples added'].values.reshape(-1, 1), values['value']).coef_.item()
                results1[feature_name] = auc
                results2[feature_name] = slope
        results1 = pd.DataFrame([results1])
        results1.insert(0, 'dataset', dataset)
        results1.insert(1, 'descriptors', descriptors)
        results1.insert(2, 'num_stds', num_stds)
        table3.append(results1)
        results2 = pd.DataFrame([results2])
        results2.insert(0, 'dataset', dataset)
        results2.insert(1, 'descriptors', descriptors)
        results2.insert(2, 'num_stds', num_stds)

table3 = pd.concat(table3)

table3.query('dataset == "All" and num_stds == 1').drop(columns=['dataset', 'num_stds']).groupby('descriptors').agg('mean').round(4).T.sort_values('biochem', ascending=False)

Obtain the second part of Figure 4D.

table5 = []
dataset  = 'All'
for descriptors in tqdm(['bio', 'biochem', 'biospectra']):
    for num_stds in range(1, 6):
        x = data.query(f'dataset == "{dataset}" and descriptors == "{descriptors}" and num_stds == {num_stds}', engine='python')
        results1 = {}
        results2 = {}
        for feature_name in chemfp_descs:
            values = x.query(f'variable == "{feature_name}"', engine='python')[['samples added', 'value']]
            if values.empty or values.shape[0] == 1:
                results1[feature_name] = None
            else:
                auc = auc_fn(values['samples added']/250, values['value'])
                slope = LinearRegression().fit(values['samples added'].values.reshape(-1, 1), values['value']).coef_.item()
                results1[feature_name] = auc
                results2[feature_name] = slope
        results1 = pd.DataFrame([results1])
        results1.insert(0, 'dataset', dataset)
        results1.insert(1, 'descriptors', descriptors)
        results1.insert(2, 'num_stds', num_stds)
        table5.append(results1)
        results2 = pd.DataFrame([results2])
        results2.insert(0, 'dataset', dataset)
        results2.insert(1, 'descriptors', descriptors)
        results2.insert(2, 'num_stds', num_stds)

table5 = pd.concat(table5)

table5.query('dataset == "All" and num_stds == 1').drop(columns=['dataset', 'num_stds']).groupby('descriptors').agg(
    'mean').round(4).T.sort_values('biochem', ascending=False)

Obtain Figure 4E.

proteins = pd.read_csv(f'{DATA_FOLDER}/05.7_combined_set_protein_targets.tsv.xz', sep='\t')
human_proteins = proteins.query('Organism == "Homo sapiens (Human)"')
human_data = data.query('variable in @human_proteins.target_id')
data_organism = data.merge(proteins[['target_id', 'Organism', 'Classification']], left_on='variable', right_on='target_id')

table11 = []
dataset  = 'All'
for descriptors in tqdm(['bio', 'biochem', 'biospectra']):
    for num_stds in range(1, 6):
        x = data_organism.query(f'dataset == "{dataset}" and descriptors == "{descriptors}" and num_stds == {num_stds}', engine='python')
        results1 = {}
        results2 = {}
        for organism in data_organism.Organism.unique():
            tmp_org_proteins = data_organism.query('Organism  == "' + organism + '"').target_id
            values = x.query(f'variable in @tmp_org_proteins', engine='python').groupby('samples added', as_index=False).sum()[['samples added', 'value']]
            if values.empty or values.shape[0] == 1:
                results1[organism] = None
            else:
                auc = auc_fn(values['samples added']/250, values['value'])
                slope = LinearRegression().fit(values['samples added'].values.reshape(-1, 1), values['value']).coef_.item()
                results1[organism] = auc
                results2[organism] = slope
        results1 = pd.DataFrame([results1])
        results1.insert(0, 'dataset', dataset)
        results1.insert(1, 'descriptors', descriptors)
        results1.insert(2, 'num_stds', num_stds)
        table11.append(results1)
        results2 = pd.DataFrame([results2])
        results2.insert(0, 'dataset', dataset)
        results2.insert(1, 'descriptors', descriptors)
        results2.insert(2, 'num_stds', num_stds)


table11 = pd.concat(table11)

table11.query('dataset == "All" and num_stds == 1').drop(columns=['dataset', 'num_stds']).groupby('descriptors').agg(
    'mean').round(4).T.sort_values('biospectra', ascending=False)

Obtain Figure 4F.

table9 = []
dataset  = 'All'
for descriptors in tqdm(['bio', 'biochem', 'biospectra']):
    for num_stds in range(1, 6):
        x = human_data.query(f'dataset == "{dataset}" and descriptors == "{descriptors}" and num_stds == {num_stds}', engine='python')
        results1 = {}
        results2 = {}
        for feature_name in human_data.variable.unique():
            values = x.query(f'variable == "{feature_name}"', engine='python')[['samples added', 'value']]
            if values.empty or values.shape[0] == 1:
                results1[feature_name] = None
            else:
                auc = auc_fn(values['samples added']/250, values['value'])
                slope = LinearRegression().fit(values['samples added'].values.reshape(-1, 1), values['value']).coef_.item()
                results1[feature_name] = auc
                results2[feature_name] = slope
        results1 = pd.DataFrame([results1])
        results1.insert(0, 'dataset', dataset)
        results1.insert(1, 'descriptors', descriptors)
        results1.insert(2, 'num_stds', num_stds)
        table9.append(results1)
        results2 = pd.DataFrame([results2])
        results2.insert(0, 'dataset', dataset)
        results2.insert(1, 'descriptors', descriptors)
        results2.insert(2, 'num_stds', num_stds)

table9 = pd.concat(table9)

table9.query('dataset == "All" and num_stds == 1').drop(columns=['dataset', 'num_stds']).groupby('descriptors').agg(
    'mean').round(4).T.sort_values('biospectra', ascending=False)

Obtain Supp. Tables 2 to 6.

ttperfs = pd.read_feather(fr'{OUT_FOLDER}/Analysis_DILI_num_samples_to_high_specificity.feather')

ttperfs.query('num_stds == 1')

ttperfs.query('num_stds == 2')

ttperfs.query('num_stds == 3')

ttperfs.query('num_stds == 4')

ttperfs.query('num_stds == 5')

Obtain Supp. Tables 7 to 10.

importances = ((tmp := data.query('num_stds == 1 and descriptors == "bio" and dataset == "All" and variable in @bio_descs'))
                           .assign(category=tmp.variable
                                               .str.split('_')
                                               .str[:3]
                                               .str.join ('_'))
                           .groupby(['category', 'samples added'])
                           .value.sum()
               )

importances_ = []
for name, group in importances.reset_index().groupby('category'):
    auc = auc_fn(group['samples added']/250, group['value'])
    importances_.append(pd.DataFrame([{'category': name, 'value': auc}]))

importances = (pd.concat(importances_, ignore_index=True)
                 .set_index('category'))
importances /= importances.value.sum()

importances = importances.reset_index()
importances = (importances.assign(reporter=importances.category.str.split('_').str[0],
                                  concentration=importances.category.str.split('_').str[1],
                                  timepoint=importances.category.str.split('_').str[2])
               ).drop(columns='category')

importances = importances.pivot(index=['reporter', 'timepoint'], columns='concentration', values='value') * 100
importances

Supp. Table 7.

importances.reset_index(level=1).iloc[:, 1:].reset_index().groupby('reporter').agg('sum').sum(axis=1).rename('total feature importance (%)').to_frame()

Supp. Table 8.

importances.sum(axis=1).rename('importnace').reset_index().pivot(index='reporter', columns='timepoint')

Supp. Table 9.

importances.reset_index(level=1).iloc[:, 1:].reset_index().groupby('reporter').agg('sum')

Obtain Supp. Table 11.

# Importances per maker type
marker_importances = ((tmp := data.query('num_stds == 1 and descriptors == "bio" and dataset == "All" and variable in @bio_descs'))
                           .assign(category=tmp.variable
                                               .str.split('_')
                                               .str[-1])
                           .groupby(['category', 'samples added'])
                           .value.sum()
               )

importances_ = []
for name, group in marker_importances.reset_index().groupby('category'):
    auc = auc_fn(group['samples added']/250, group['value'])
    importances_.append(pd.DataFrame([{'category': name, 'value': auc}]))

marker_importances = (pd.concat(importances_, ignore_index=True)
                 .set_index('category'))
marker_importances /= marker_importances.value.sum()

marker_importances = marker_importances.reset_index()
marker_importances
